requirements.txt
bcrypt==4.3.0
beautifulsoup4==4.13.4
pillow==11.2.1
requests==2.32.3